+++ cmd/webserver/game.db.json
@@ -1,0 +1,1 @@
+[]
