+++ server.go
@@ -1,4 +1,4 @@
-package main
+package poker
 
 import (
 	"encoding/json"
